+++ src/media-popup.html
@@ -10,5 +10,6 @@
       <img id="loadedImage" src=></img>
       <video id="loadedVideo" src=></video>
     </div>
+    <script src="media-popup.js"></script>
   </body>
 </html>
